% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny.R
\name{create.shiny}
\alias{create.shiny}
\title{Generate shiny interface for visualization and exploration}
\usage{
create.shiny(tsne_dt)
}
\arguments{
\item{object}{t-Sne clustering object that has been created.}
}
\description{
This function generate an interactive shiny application for visualizing and exploring treatment effect estimated by GRF.
}
\examples{
create.shiny(tsne_obj$result)
}
