% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny.R
\name{create_shiny}
\alias{create_shiny}
\title{Generate shiny interface for data visualization and exploration}
\usage{
create_shiny(tsne_dt)
}
\arguments{
\item{object}{t-Sne clustering object that has been created.}
}
\description{
This function generate an interactive shiny application that can be used to visualize and explore heterogeneous effect captured by GRF.
}
\examples{
create_shiny(tsne_obj$result)
}
