% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohieData.R
\docType{data}
\name{ohieData}
\alias{ohieData}
\title{Survey respondents dataset from Oregon Health Insurance Experiment (OHIE)}
\format{An object of class \code{data.frame} with 23741 rows and 227 columns.}
\source{
\url{http://www.nber.org/oregon/4.data.html}
}
\usage{
ohieData
}
\description{
The original survey data consists 58405 survey participants in total,
including both 23714 respondents and 51184 non-respondents.
To keep congruent with Finkelstein’s analysis
this dataset only consists complete cases, which regards 23714 survey respondents.
}
\keyword{datasets}
