% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cf.R
\name{read_cf}
\alias{read_cf}
\title{Read causal forest object}
\usage{
read_cf(object, X = "All")
}
\arguments{
\item{object}{causal forest object that has been created.}

\item{X}{Vector of covariate names or number of clumns to includes in later t-SNE analysis.Defaults to all covariates in X.orig of cf object.}
}
\value{
A dataframe with selected covariates that is ready for t-SNE clustering analysis
}
\description{
tSNE-Clustering visualization start by a causal forest object. 
This function read a causal forest object that has been created and 
process it for next t-SNE clustering analysis.
}
\examples{
# Train a causal forest.
n = 50; p = 10
X = matrix(rnorm(n*p), n, p)
W = rbinom(n, 1, 0.5)
Y = pmax(X[,1], 0) * W + X[,2] + pmin(X[,3], 0) + rnorm(n)
c.forest <- causal_forest(X, Y, W)
my_cf <- read_cf(c.forest, X = c(2:7))
}
\references{
Athey, Susan, Julie Tibshirani, and Stefan Wager. Generalized Random Forests. Annals of Statistics (forthcoming), 2018

Wager, Stefan, and Susan Athey. Estimation and Inference of Heterogeneous Treatment Effects using Random Forests. Journal of the American Statistical Association (forthcoming), 2018.
}
