% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run.analsis}
\alias{run.analsis}
\title{Run t-SNE clustering algorithm}
\usage{
run.analsis(object, n, distance = NULL, perp = NULL)
}
\arguments{
\item{object}{causal forest object that has been created.}

\item{n}{size of subsample from whole dataset. Suggest no larger than 10,000.}

\item{distance}{distance metrix of covariate space either provided by user or use default: gower distance}

\item{perplexity}{Perplexity parameter (should no larger than 3 * perplexity < nrow(X) - 1, see details for interpretation).}
}
\value{
A t-SNE clustering object with the follwing elements:
\item{distance} distance metrix of covariate space.
\item{result} final datafram proceed for visualization
}
\description{
This function read previous proceeded causal forest object and get 2D projection of covariates space based on t-SNE algorithm
}
\examples{
tsne_obj <- run.analsis(my_cf)
}
\references{
Maaten, L. Van Der, 2014. Accelerating t-SNE using Tree-Based Algorithms. Journal of Machine Learning Research, 15, p.3221-3245

van der Maaten, L.J.P. & Hinton, G.E., 2008. Visualizing High-Dimensional Data Using t-SNE. Journal of Machine Learning Research, 9, pp.2579-2605.
}
